RE2_VM_NFA::get_non_epsilon_targets() del old comment
The old comment is wrong, it is misleading.

re2::Prog::Inst::foldcase() is not controled by
Regexp::ParseFlags::FoldCase, it is used to optimize
regex like [A-Za-z] into one Instruction(Prog::Inst),
if Inst.foldcase is true, upper case is also included.

This is why the 2 prev commit is needed.